# IJICommerce/api-documentation.yaml
openapi: 3.0.0
info:
  title: IJICommerce API (Phase 1 & Early Phase 2)
  version: v0.2.0
  description: API for managing Shops, Shop Teams, Product Proposals, Brands, and Categories.
  contact:
    name: IJIDeals
    email: contact@ijideals.com
servers:
  - url: http://localhost/api/ijicommerce # Replace with your actual base URL + package route_prefix
    description: Local development server
components:
  securitySchemes:
    bearerAuth: # Can be named anything
      type: http
      scheme: bearer
      bearerFormat: Sanctum # Or JWT, etc.
security:
  - bearerAuth: []
paths:
...
  /shops/{shopId}:
    get:
      summary: Get a specific shop's details
      tags:
        - Shops
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: path
          required: true
          description: ID of the shop
          schema:
            type: integer
      responses:
        '200':
          description: Shop details.
        '403':
          description: Unauthorized to view this shop.
        '404':
          description: Shop not found.
    put:
      summary: Update a specific shop
      tags:
        - Shops
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                # ... other updatable shop fields ...
      responses:
        '200':
          description: Shop updated successfully.
        '403':
          description: Unauthorized to update this shop.
    delete:
      summary: Delete a specific shop
      tags:
        - Shops
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop deleted successfully.
        '403':
          description: Unauthorized to delete this shop.
...
  /shops/{shopId}/shop-products/{shopProductId}:
    get:
      summary: Get a specific shop product listing
      tags:
        - Shop Product Listings
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: shopProductId # Corrected from shopProduct
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop product listing details.
    put:
      summary: Update a shop product listing
      tags:
        - Shop Product Listings
      # ... parameters, requestBody, responses ...
      parameters: []
      requestBody: {}
      responses: {}
    delete:
      summary: Remove a product listing from the shop
      tags:
        - Shop Product Listings
      # ... parameters, responses ...
      parameters: []
      responses: {}
...